Removed the middleware folder and moved the contents to a middleware file in the controllers folder
