+++ internal/controllers/middleware.go
@@ -1,6 +1,7 @@
-package middleware
+package controllers
 
 import (
+	"net/http"
 	"strings"
 
 	"github.com/gin-gonic/gin"
@@ -38,4 +39,14 @@ func parseError(err error) (string, int) {
 	}
 
 	return errMessage, code
+}
+
+func CustomRecoveryMiddleware() func(c *gin.Context, recovered interface{}) {
+	return func(c *gin.Context, recovered interface{}) {
+		c.JSON(500, gin.H{
+			"status":  500,
+			"message": "Internal Server Error",
+		})
+		c.AbortWithStatus(http.StatusInternalServerError)
+	}
 }
